pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "End-to-End Q&A Chatbot using LangChain with Groq, Gemini, HuggingFace, Web Search, RAG, and Voice"
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
    "streamlit>=1.34.0",
    "langchain>=0.3.26",
    "langchain-openai>=0.1.7",
    "langchain-community>=0.0.38",
    "langchain-core>=0.1.60",
    "langchain-groq>=0.3.8",
    "langchain-text-splitters>=0.3.0",
    "langchain-google-genai>=0.0.11",
    "langchain-huggingface>=0.1.0",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.8.3",
    "google-search-results>=2.4.3",
    "gtts>=2.5.1",
    "SpeechRecognition>=3.10.1",
    "faiss-cpu>=1.8.0",
    "docx2txt>=0.8",
    "pypdf>=4.0.2",
    "sentence-transformers>=2.2.2",
    "tiktoken>=0.7.0",
    "psutil>=5.9.0",
    "python-pptx>=0.6.23",
    "Pillow>=10.0.0",
]
